divided in functions

+++ view_item.php
@@ -10,20 +10,31 @@ include "head.php";
         <label for="sel_followed">Rechercher animal selon:</label>
         <select name="search_field" id="sel_followed" class="form-control">
 <?php //Creates list of choices
-$columns = get_columns('Followed');
-$displayed_fields = array();  //field => displayed
-$keys_tables = main_tables_from_keys();
-foreach ($columns as $col_specs)
+function create_fields_array($columns)
 {
-    $field = $col_specs['Field'];
-    if ($col_specs['Key'] == 'MUL')
+    $displayed_fields = array();  //field => displayed
+    $keys_tables = main_tables_from_keys();
+    foreach ($columns as $col_specs)
     {
-        $displayed_fields[$field] = $keys_tables[$field];
+        $field = $col_specs['Field'];
+        if ($col_specs['Key'] == 'MUL')
+        {
+            $displayed_fields[$field] = $keys_tables[$field];
+        }
+        else $displayed_fields[$field] = $field;
     }
-    else $displayed_fields[$field] = $field;
-    echo "<option value=\"$field\">" . ucfirst($displayed_fields[$field]) .
-        "</option>";
+    return $displayed_fields;
 }
+
+function create_choice_list($displayed_fields)
+{
+    foreach ($displayed_fields as $field => $disp)
+    {
+        echo "<option value=\"$field\">" . ucfirst($disp) . "</option>";
+    }
+}
+$columns = get_columns('Followed');
+$displayed_fields = create_fields_array($columns);
 ?>
         </select>
     </div>
@@ -31,11 +42,11 @@ foreach ($columns as $col_specs)
 </form>
 
 <?php //Creates array
-// Create first row containing fields
-if (array_key_exists('search_field', $_POST))
+function create_tablehead($search_field, $columns, $displayed_fields)
 {
-    $search_field = $_POST['search_field'];
-    echo '<table class="table" data-toggle="table" data-search="true">';
+    //$search_field must be a columns name
+    //$columns result of SHOW columns FROM
+    //$displayed fields array ($col_name => $displayed_value
     echo '<thead><tr>';
     echo '<th data-filed="s_field" data-sortable="true">' .
         ucfirst($search_field) . '</th>';
@@ -55,6 +66,12 @@ if (array_key_exists('search_field', $_POST))
     echo '</tr></thead>';
 
     // Create body containing results
+}
+// Create first row containing fields
+if (array_key_exists('search_field', $_POST))
+{
+    echo '<table class="table" data-toggle="table" data-search="true">';
+    create_tablehead($_POST['search_field'], $columns, $displayed_fields);
     echo '<tbody>';
     echo '</tbody>';
     echo '</table>';
